% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calibrateEB.R
\name{calibrateEB}
\alias{calibrateEB}
\title{Empirical Bayes calibration of noisy variance estimates}
\usage{
calibrateEB(vars, sigma2)
}
\arguments{
\item{vars}{list of variance estimates}

\item{sigma2}{estimate of the Monte Carlo noise in vars}
}
\value{
calibrated variance estimates
}
\description{
Empirical Bayes calibration of noisy variance estimates
}

