% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calibrateEB.R
\name{gbayes}
\alias{gbayes}
\title{Bayes posterior estimation with Gaussian noise}
\usage{
gbayes(x0, g.est, sigma)
}
\arguments{
\item{x0}{an obsevation}

\item{g.est}{a prior density, as returned by gfit}

\item{sigma}{noise estimate}
}
\value{
posterior estimate E[mu | x0]
}
\description{
Bayes posterior estimation with Gaussian noise
}

