% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calibrateEB.R
\name{gfit}
\alias{gfit}
\title{Fit an empirical Bayes prior in the hierarchical model
    mu ~ G, X ~ N(mu, sigma^2)}
\usage{
gfit(X, sigma, p = 2, nbin = 1000, unif.fraction = 0.1)
}
\arguments{
\item{X}{a vector of observations}

\item{sigma}{noise estimate}

\item{p}{tuning parameter -- number of parameters used to fit G}

\item{nbin}{tuning parameter -- number of bins used for discrete approximation}

\item{unif.fraction}{tuning parameter -- fraction of G modeled as "slab"}
}
\value{
posterior density estimate g
}
\description{
Fit an empirical Bayes prior in the hierarchical model
    mu ~ G, X ~ N(mu, sigma^2)
}
\section{References}{

For more details about "g-estimation", see: B Efron. Two modeling strategies for
empirical Bayes estimation. Stat. Sci., 29(2): 285–301, 2014.
}

