% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinitesimalJackknife.R
\name{infJack}
\alias{infJack}
\title{The infinitesimal jackknife for random forests}
\usage{
infJack(pred, inbag, calibrate = TRUE, used.trees = NULL)
}
\arguments{
\item{pred}{A nrow(newdata) by no. of trees matrix which contains numeric predictions
from a random forest trained with trees grown on bootstrap samples of the training data}

\item{inbag}{A number of obs. in the training data by no. of trees matrix giving the
number of times the ith observation in the training data appeared in the bootstrap sample for the jth tree.}

\item{calibrate}{whether to apply calibration to mitigate Monte Carlo noise
warning: if calibrate = FALSE, some variance estimates may be negative
         due to Monte Carlo effects if the number of trees in rf is too small}

\item{used.trees}{set of trees to use for variance estimation; uses all tress if NULL}
}
\description{
The infinitesimal jackknife for random forests
}
