% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/infinitesimalJackknife.R
\name{randomForestInfJackSubsample}
\alias{randomForestInfJackSubsample}
\title{The infinitesimal jackknife for random forests with subsampling}
\usage{
randomForestInfJackSubsample(rf, newdata)
}
\arguments{
\item{rf}{A random forest trained with replace = TRUE and keep.inbag = TRUE}

\item{newdata}{A set of test points at which to evaluate standard errors}
}
\description{
The infinitesimal jackknife for random forests with subsampling
}

