% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/infinitesimalJackknife.R
\name{randomForestInfJack}
\alias{randomForestInfJack}
\title{The infinitesimal jackknife for random forests}
\usage{
randomForestInfJack(rf, newdata)
}
\arguments{
\item{rf}{A random forest trained with replace = TRUE and keep.inbag = TRUE}

\item{newdata}{A set of test points at which to evaluate standard errors}
}
\description{
The infinitesimal jackknife for random forests
}

